+++ tests/serve/test_vllm.py
@@ -14,12 +14,8 @@ from tests.serve.common import (
     params_with_model_mark,
     run_serve_deployment,
 )
-from tests.serve.conftest import (
-    MULTIMODAL_IMG_PATH,
-    MULTIMODAL_IMG_URL,
-    MinioLoraConfig,
-    load_lora_adapter,
-)
+from tests.serve.conftest import MULTIMODAL_IMG_PATH, MULTIMODAL_IMG_URL
+from tests.serve.lora_utils import MinioLoraConfig, load_lora_adapter
 from tests.utils.engine_process import EngineConfig
 from tests.utils.payload_builder import (
     chat_payload,
